(cleanup) Relocate files and update file paths
